requirements.txt
rumps>=0.4.0
noisereduce>=3.0.0
soundfile>=0.12.0
watchdog>=3.0.0
openai-whisper>=20231117
mlx-whisper>=0.4.3
lightning-whisper-mlx>=0.0.10
mlx>=0.31.0
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
torch>=2.0.0
py2app>=0.28.0
setproctitle>=1.3.0
pyobjc-framework-Quartz>=9.0
pyobjc-framework-AVFoundation>=9.0
pyobjc-framework-ScreenCaptureKit>=12.0
